+++ src/utils/config-manager.ts
@@ -6,7 +6,7 @@ import { isConfig, isSettings } from '../types';
 import type { Config, Settings, CLIProxyVariantsConfig, CLIProxyVariantConfig } from '../types';
 import { expandPath, error } from './helpers';
 import { info } from './ui';
-import { isUnifiedMode, loadOrCreateUnifiedConfig } from '../config/config-loader-facade';
+import { isUnifiedMode, loadOrCreateUnifiedConfig } from '../config/unified-config-loader';
 
 // TODO: Replace with proper imports after converting these files
 // const { ErrorManager } = require('./error-manager');
